+++ cmd/cryptool/root/subcmds/coinbase_wallet.go
@@ -1,0 +1,12 @@
+package subcmds
+
+import "github.com/spf13/cobra"
+
+func newCoinbaseWalletCmd() *cobra.Command {
+	cmd := &cobra.Command{
+		Use:   "wallet",
+		Short: "Wallet related commands for Coinbase",
+	}
+	cmd.AddCommand(newCoinbaseWalletSyncDownCmd())
+	return cmd
+}